Improve POS register delivery flow
