+++ .graphify_python
@@ -1,1 +1,1 @@
-/opt/homebrew/bin/python3
+/opt/homebrew/bin/python3.10
