+++ packages/athena-webapp/src/lib/pos/presentation/registerSessionCode.ts
@@ -1,0 +1,9 @@
+export function formatRegisterSessionCode(sessionId?: string | null) {
+  const trimmedSessionId = sessionId?.trim();
+
+  if (!trimmedSessionId) {
+    return undefined;
+  }
+
+  return trimmedSessionId.slice(-6).toUpperCase();
+}
