+++ packages/athena-webapp/src/routes/login/-login-ready-view.tsx
@@ -1,0 +1,9 @@
+import { Login } from "~/src/components/auth/Login";
+
+export function AthenaLoginReadyView() {
+  return (
+    <section data-testid="athena-login-ready">
+      <Login />
+    </section>
+  );
+}
