+++ packages/athena-webapp/src/routes/login/_layout.index.tsx
@@ -1,13 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { Login } from "~/src/components/auth/Login";
-
-export function AthenaLoginReadyView() {
-  return (
-    <section data-testid="athena-login-ready">
-      <Login />
-    </section>
-  );
-}
+import { AthenaLoginReadyView } from "./-login-ready-view";
 
 export const Route = createFileRoute("/login/_layout/")({
   component: AthenaLoginReadyView,
